requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
huggingface-hub==0.19.4
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1requests==2.31.0